[ANDROSDK-1438] Refactor AnalyticsException package
