+++ core/src/main/java/org/hisp/dhis/android/core/analytics/AnalyticsException.kt
@@ -26,7 +26,7 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.analytics.aggregated.internal
+package org.hisp.dhis.android.core.analytics
 
 import org.hisp.dhis.android.core.program.ProgramIndicator
