+++ core/src/main/java/org/hisp/dhis/android/core/analytics/aggregated/internal/AnalyticsService.kt
@@ -29,6 +29,7 @@
 package org.hisp.dhis.android.core.analytics.aggregated.internal
 
 import android.database.sqlite.SQLiteException
+import org.hisp.dhis.android.core.analytics.AnalyticsException
 import javax.inject.Inject
 import org.hisp.dhis.android.core.analytics.aggregated.Dimension
 import org.hisp.dhis.android.core.analytics.aggregated.DimensionalResponse
