+++ core/src/main/java/org/hisp/dhis/android/core/analytics/aggregated/internal/evaluator/IndicatorEvaluator.kt
@@ -30,7 +30,7 @@ package org.hisp.dhis.android.core.analytics.aggregated.internal.evaluator
 import javax.inject.Inject
 import org.hisp.dhis.android.core.analytics.aggregated.DimensionItem
 import org.hisp.dhis.android.core.analytics.aggregated.MetadataItem
-import org.hisp.dhis.android.core.analytics.aggregated.internal.AnalyticsException
+import org.hisp.dhis.android.core.analytics.AnalyticsException
 import org.hisp.dhis.android.core.analytics.aggregated.internal.AnalyticsServiceEvaluationItem
 import org.hisp.dhis.android.core.analytics.aggregated.internal.evaluator.indicatorengine.IndicatorEngine
 import org.hisp.dhis.android.core.indicator.Indicator
